arno-pinyin/backend/requirements.txt
flask
flask-cors
pyjwt
python-openai
sqlite3
